pyproject.toml
[project]
name = "fastmcp-gsuite"
version = "0.4.1"
description = "MCP Server to connect to Google G-Suite"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "beautifulsoup4>=4.12.3",
 "google-api-python-client>=2.154.0",
 "httplib2>=0.22.0",
 # "mcp>=1.3.0", # Now included via fastmcp
 "oauth2client==4.1.3",
 "python-dotenv>=1.0.1",
 "pytz>=2024.2",
 "requests>=2.32.3",
 "fastmcp>=2.1.2,<2.2.0",
 "pydantic-settings>=2.8.1,<2.9.0",
]
[[project.authors]]
name = "Markus Pfundstein"
email = "markus@life-electronic.nl"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gsuite"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.5",
    "httpx>=0.27.0",
    "google-auth-oauthlib>=1.2.1",
    "pytest-asyncio>=0.23.0",
    "fastmcp-gsuite[e2e]",
]

e2e = [
    "pytest>=8.3.5",
    "google-auth>=2.27.0",
    "google-api-python-client>=2.154.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.389",
]
e2e = [
    "pytest>=8.3.5", 
    "google-auth>=2.27.0",
    "google-api-python-client>=2.154.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.2.1",
]

[project.scripts]
# mcp-gsuite = "mcp_gsuite:main" # Old entry point
fastmcp-gsuite-fast = "mcp_gsuite.fast_server:mcp.run" # New entry point using fastmcp
fastmcp-gsuite = "mcp_gsuite.fast_server:mcp.run" # Alias for fastmcp implementation

# Add pytest configuration to include project root in pythonpath
[tool.pytest.ini_options]
pythonpath = [
  "."
]
markers = [
  "e2e: marks tests as end-to-end tests requiring real credentials"
]

[tool.uv.sources]
fastmcp-gsuite = { workspace = true }
